Clean BatteryBar - move to ui/layout - fix incorrect battery color handling
