+++ src/ui/layout/BatteryBar.vue
@@ -4,14 +4,27 @@ import { useBatteryLevelStore } from '@/stores'
 import colors from 'vuetify/lib/util/colors'
 
 const batteryLevelStore = useBatteryLevelStore()
-const coverLevel = computed(() => 100 - batteryLevelStore.percentage + '%')
+const coverLevel = computed(() => {
+    return 100 - batteryLevelStore.percentage + '%'
+})
 
 const batteryColor = computed(() => {
-    console.log(batteryLevelStore.percentage)
-    if (batteryLevelStore.percentage < 75) return colors.green.base
-    if (batteryLevelStore.percentage < 50) return colors.amber.base
-    if (batteryLevelStore.percentage < 25) return colors.red.base
-    if (batteryLevelStore.percentage < 10) return colors.pink.base
+    if (batteryLevelStore.percentage < 10) {
+        return colors.pink.base
+    }
+
+    if (batteryLevelStore.percentage < 25) {
+        return colors.red.base
+    }
+
+    if (batteryLevelStore.percentage < 50) {
+        return colors.amber.base
+    }
+
+    if (batteryLevelStore.percentage < 75) {
+        return colors.green.base
+    }
+
     return colors.teal.lighten2
 })
 const batteryBackgroundColor = colors.blueGrey.darken2
@@ -34,9 +47,9 @@ const batteryBackgroundColor = colors.blueGrey.darken2
     background-color: v-bind(batteryColor);
 }
 .battery-level {
+    width: 100%;
     height: v-bind(coverLevel);
     transition: height 0.2s cubic-bezier(0.4, 0, 0.2, 1);
-    width: 100%;
     background-color: v-bind(batteryBackgroundColor);
 }
 </style>
